requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
arrow==1.3.0
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==4.2.0
blinker==1.7.0
bs4==0.0.1
cachetools==5.3.2
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
cryptography==41.0.7
decorator==5.1.1
distro==1.9.0
dj-database-url==2.1.0
Django==4.2.8
django-appconf==1.0.6
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-cryptography==1.1
django-fernet-fields==0.6
django-heroku==0.3.1
django-timezone-field==6.1.0
django-utils==0.0.2
djangorestframework==3.14.0
drf-spectacular
et-xmlfile==1.1.0
frozenlist==1.4.1
fsspec==2023.12.2
gcsfs==2023.12.2.post1
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.15.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
importlib-metadata==6.11.0
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kombu==5.3.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
mygeotab==0.8.8
numpy==1.26.2
oauthlib==3.2.2
openai==1.6.1
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
prompt-toolkit==3.0.43
protobuf==4.25.1
psycopg2==2.9.9
pyarrow==14.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydeck==0.8.1b0
Pygments==2.17.2
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-rapidjson==1.14
pytz==2023.3.post1
redis==5.0.1
referencing==0.32.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.15.2
rsa==4.9
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
streamlit==1.29.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
types-python-dateutil==2.8.19.14
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
vine==5.1.0
wcwidth==0.2.12
whitenoise==6.6.0
yarl==1.9.4
zipp==3.17.0